+++ NEWS.md
@@ -1,3 +1,13 @@
+# evabic 0.0.3
+
+* Functions error if logical `detected` or `true` do not have names. 
+
+# evabic 0.0.2
+
+* New function `ebc_AUC_from_measures()` to compute AUC directly from the tidy dataframe.
+
+* New function `add_names()` to add (eventually default) names to a vector.
+
 # evabic 0.0.1
 
 * First release.